# docs/swagger.yaml
definitions:
  model.Brand:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        description: 品牌名称
        type: string
    type: object
  model.Build:
    properties:
      bakingtem:
        description: 水重量g
        type: number
      bakingtime:
        description: 水重量g
        type: number
      build_data:
        allOf:
        - $ref: '#/definitions/model.BuildData'
        description: 字典 材料id-使用克重
      created_at:
        type: string
      id:
        type: integer
      info:
        description: 制作流程
        type: string
      materialUnit:
        description: 水重量g
        type: number
      name:
        description: 工艺名称
        type: string
      remake:
        description: 备注
        type: string
      water:
        description: 水重量g
        type: number
    type: object
  model.BuildData:
    additionalProperties:
      type: number
    type: object
  model.Goods:
    properties:
      barcode:
        description: 产品条码
        type: string
      brand:
        description: 品牌
        type: string
      build_id:
        description: 关联制作工艺
        type: integer
      created_at:
        type: string
      id:
        description: 产品ID
        type: integer
      name:
        description: 产品名称
        type: string
      price:
        description: 产品价格
        type: number
      ptype:
        $ref: '#/definitions/model.Goodstype'
      remake:
        description: 备注
        type: string
      unit:
        description: 产品规格/g
        type: number
    type: object
  model.Goodstype:
    enum:
    - 成品
    - 原料
    type: string
    x-enum-varnames:
    - Product
    - Material
  request.GetById:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  request.RequestInventory:
    properties:
      id:
        description: 货品id
        type: integer
      num:
        type: integer
    type: object
  request.RequestProduction:
    properties:
      gid:
        additionalProperties:
          type: integer
        description: 货品id 数量
        type: object
    type: object
  request.SearchBrandParams:
    properties:
      created_at:
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      id:
        type: integer
      keyword:
        description: 关键字
        type: string
      name:
        description: 品牌名称
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
  request.SearchGoodsParams:
    properties:
      barcode:
        description: 产品条码
        type: string
      brand:
        description: 品牌
        type: string
      build_id:
        description: 关联制作工艺
        type: integer
      created_at:
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      id:
        description: 产品ID
        type: integer
      keyword:
        description: 关键字
        type: string
      name:
        description: 产品名称
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      price:
        description: 产品价格
        type: number
      ptype:
        $ref: '#/definitions/model.Goodstype'
      remake:
        description: 备注
        type: string
      unit:
        description: 产品规格/g
        type: number
    type: object
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: gin+vue
  title: 进销存 Swagger API接口文档
  version: v0.0.1
paths:
  /api/DeleteGoodsByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除
      tags:
      - FactoryApi
  /api/deleteBrandByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除选中品牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除选中brand
      tags:
      - FactoryApi
  /api/deleteBuildByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除选中brand
      tags:
      - FactoryApi
  /api/showCost:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: 展示成本
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改基础api
      tags:
      - FactoryApi
  /api/updateBrand:
    post:
      consumes:
      - application/json
      parameters:
      - description: 品牌信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Brand'
      produces:
      - application/json
      responses:
        "200":
          description: 修改品牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改品牌
      tags:
      - FactoryApi
  /api/updateBuild:
    post:
      consumes:
      - application/json
      parameters:
      - description: 工艺信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Build'
      produces:
      - application/json
      responses:
        "200":
          description: 修改品牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改基础api
      tags:
      - FactoryApi
  /api/updateGoods:
    post:
      consumes:
      - application/json
      parameters:
      - description: 产品信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Goods'
      produces:
      - application/json
      responses:
        "200":
          description: 修改产品
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改基础api
      tags:
      - FactoryApi
  /factory/createBrand:
    post:
      consumes:
      - application/json
      parameters:
      - description: 品牌信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Brand'
      produces:
      - application/json
      responses:
        "200":
          description: 创建品牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建品牌
      tags:
      - FactoryApi
  /factory/createBuild:
    post:
      consumes:
      - application/json
      parameters:
      - description: 工艺信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Build'
      produces:
      - application/json
      responses:
        "200":
          description: 创建工艺流程
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建工艺流程
      tags:
      - FactoryApi
  /factory/createGoods:
    post:
      consumes:
      - application/json
      parameters:
      - description: 产品信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Goods'
      produces:
      - application/json
      responses:
        "200":
          description: 创建产品
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建产品
      tags:
      - FactoryApi
  /factory/getBrandList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页获取品牌列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchBrandParams'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取品牌列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取品牌列表
      tags:
      - FactoryApi
  /factory/getBuildList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页获取品牌列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchBrandParams'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取品牌列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取工艺信息
      tags:
      - FactoryApi
  /factory/getGoodsList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 产品列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchGoodsParams'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取产品列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取prduct
      tags:
      - FactoryApi
  /factory/production:
    post:
      consumes:
      - application/json
      parameters:
      - description: 库存
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RequestInventory'
      produces:
      - application/json
      responses:
        "200":
          description: 生产
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 进出货
      tags:
      - FactoryApi
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"